+++ src/core/event_registry/server_event.rs
@@ -340,9 +340,9 @@ impl ServerEvent {
         events: PtrMut,
         queue: PtrMut,
         client: &mut RepliconClient,
-        init_tick: RepliconTick,
+        change_tick: RepliconTick,
     ) {
-        (self.receive)(self, ctx, events, queue, client, init_tick);
+        (self.receive)(self, ctx, events, queue, client, change_tick);
     }
 
     /// Drains events [`ToClients<E>`] and re-emits them as `E` if the server is in the list of the event recipients.
@@ -494,12 +494,12 @@ unsafe fn receive<E: Event>(
     events: PtrMut,
     queue: PtrMut,
     client: &mut RepliconClient,
-    init_tick: RepliconTick,
+    change_tick: RepliconTick,
 ) {
     let events: &mut Events<E> = events.deref_mut();
     let queue: &mut ServerEventQueue<E> = queue.deref_mut();
 
-    while let Some((tick, message)) = queue.pop_if_le(init_tick) {
+    while let Some((tick, message)) = queue.pop_if_le(change_tick) {
         let mut cursor = Cursor::new(&*message);
         match event_data.deserialize(ctx, &mut cursor) {
             Ok(event) => {
@@ -529,7 +529,7 @@ unsafe fn receive<E: Event>(
                     continue;
                 }
             };
-            if tick > init_tick {
+            if tick > change_tick {
                 trace!("queuing event `{}` with `{tick:?}`", any::type_name::<E>());
                 queue.insert(tick, message.slice(cursor.position() as usize..));
                 continue;
@@ -699,19 +699,19 @@ enum SerializedMessage {
 }
 
 impl SerializedMessage {
-    /// Optimized to avoid reallocations when clients have the same init tick as other clients receiving the
+    /// Optimized to avoid reallocations when clients have the same change tick as other clients receiving the
     /// same message.
-    fn get_bytes(&mut self, init_tick: RepliconTick) -> bincode::Result<Bytes> {
+    fn get_bytes(&mut self, change_tick: RepliconTick) -> bincode::Result<Bytes> {
         match self {
             // Resolve the raw value into a message with serialized tick.
             Self::Raw(raw) => {
                 let mut bytes = std::mem::take(raw);
-                let tick_size = DefaultOptions::new().serialized_size(&init_tick)? as usize;
+                let tick_size = DefaultOptions::new().serialized_size(&change_tick)? as usize;
                 let padding = RepliconTick::MAX_SERIALIZED_SIZE - tick_size;
-                DefaultOptions::new().serialize_into(&mut bytes[padding..], &init_tick)?;
+                DefaultOptions::new().serialize_into(&mut bytes[padding..], &change_tick)?;
                 let bytes = Bytes::from(bytes).slice(padding..);
                 *self = Self::Resolved {
-                    tick: init_tick,
+                    tick: change_tick,
                     tick_size,
                     bytes: bytes.clone(),
                 };
@@ -723,13 +723,13 @@ impl SerializedMessage {
                 tick_size,
                 bytes,
             } => {
-                if *tick == init_tick {
+                if *tick == change_tick {
                     return Ok(bytes.clone());
                 }
 
-                let new_tick_size = DefaultOptions::new().serialized_size(&init_tick)? as usize;
+                let new_tick_size = DefaultOptions::new().serialized_size(&change_tick)? as usize;
                 let mut new_bytes = Vec::with_capacity(new_tick_size + bytes.len() - *tick_size);
-                DefaultOptions::new().serialize_into(&mut new_bytes, &init_tick)?;
+                DefaultOptions::new().serialize_into(&mut new_bytes, &change_tick)?;
                 new_bytes.extend_from_slice(&bytes[*tick_size..]);
                 Ok(new_bytes.into())
             }
@@ -749,7 +749,7 @@ impl BufferedServerEvent {
         server: &mut RepliconServer,
         client: &ReplicatedClient,
     ) -> bincode::Result<()> {
-        let message = self.message.get_bytes(client.init_tick())?;
+        let message = self.message.get_bytes(client.change_tick())?;
         server.send(client.id(), self.channel, message);
         Ok(())
     }
@@ -769,10 +769,10 @@ impl BufferedServerEventSet {
     }
 }
 
-/// Caches synchronization-dependent server events until they can be sent with an accurate init tick.
+/// Caches synchronization-dependent server events until they can be sent with an accurate change tick.
 ///
 /// This exists because replication does not scan the world every tick. If a server event is sent in the same
-/// tick as a spawn and the event references that spawn, then the server event's init tick needs to be synchronized
+/// tick as a spawn and the event references that spawn, then the server event's change tick needs to be synchronized
 /// with that spawn on the client. We buffer the event until the spawn can be detected.
 #[derive(Resource, Default)]
 pub(crate) struct BufferedServerEvents {
@@ -890,9 +890,9 @@ struct ServerEventQueue<E> {
 
 impl<E> ServerEventQueue<E> {
     /// Pops the next event that is at least as old as the specified replicon tick.
-    fn pop_if_le(&mut self, init_tick: RepliconTick) -> Option<(RepliconTick, Bytes)> {
+    fn pop_if_le(&mut self, change_tick: RepliconTick) -> Option<(RepliconTick, Bytes)> {
         let (tick, _) = self.list.front()?;
-        if *tick > init_tick {
+        if *tick > change_tick {
             return None;
         }
         self.list