+++ src/server/replication_messages.rs
@@ -1,8 +1,8 @@
-pub(super) mod init_message;
+pub(super) mod change_message;
 pub(super) mod mutate_message;
 pub(super) mod serialized_data;
 
-use init_message::InitMessage;
+use change_message::ChangeMessage;
 use mutate_message::MutateMessage;
 
 /// Accumulates replication messages.
@@ -12,7 +12,7 @@ use mutate_message::MutateMessage;
 /// serialized data across messages.
 #[derive(Default)]
 pub(crate) struct ReplicationMessages {
-    messages: Vec<(InitMessage, MutateMessage)>,
+    messages: Vec<(ChangeMessage, MutateMessage)>,
     len: usize,
 }
 
@@ -30,8 +30,8 @@ impl ReplicationMessages {
         self.messages.reserve(additional);
 
         for index in 0..clients_count {
-            if let Some((init_message, mutate_message)) = self.messages.get_mut(index) {
-                init_message.clear();
+            if let Some((change_message, mutate_message)) = self.messages.get_mut(index) {
+                change_message.clear();
                 mutate_message.clear();
             } else {
                 self.messages.push(Default::default());
@@ -40,7 +40,7 @@ impl ReplicationMessages {
     }
 
     /// Returns iterator over messages for each client.
-    pub(super) fn iter_mut(&mut self) -> impl Iterator<Item = &mut (InitMessage, MutateMessage)> {
+    pub(super) fn iter_mut(&mut self) -> impl Iterator<Item = &mut (ChangeMessage, MutateMessage)> {
         self.messages.iter_mut().take(self.len)
     }
 }