+++ src/server/replication_messages/serialized_data.rs
@@ -16,7 +16,7 @@ use crate::{
 
 /// Single continious buffer that stores serialized data for messages.
 ///
-/// See [`InitMessage`](super::init_message::InitMessage) and
+/// See [`ChangeMessage`](super::change_message::ChangeMessage) and
 /// [`MutateMessage`](super::mutate_message::MutateMessage).
 #[derive(Default, Deref, DerefMut)]
 pub(crate) struct SerializedData(Vec<u8>);