+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move `Replicated` to the `replication` module.
 - Split the `ctx` module and move event-related contexts under `core::events_registry::ctx` and replication-related contexts under `core::replication_registry::ctx`.
 - Rename `ServerPlugin::change_timeout` into `ServerPlugin::mutate_timeout`.
+- Rename `ServerInitTick` into `ServerChangeTick`.
+- Rename `ReplicatedClient::init_tick` into `ReplicatedClient::change_tick`.
+- Rename `ReplicatedClient::get_change_tick` into `ReplicatedClient::mutation_tick`.
+- Rename `ReplicationChannel::Init` into `ReplicationChannel::Changes`.
 - Rename `ReplicationChannel::Update` into `ReplicationChannel::Mutations`.
 
 ### Removed